Extract more formatting code from DPVGenerator to DPVWriter.
