+++ src/test/groovy/com/seriouslypro/pnpconvert/DPVWriterSpec.groovy
@@ -1,23 +1,27 @@
 package com.seriouslypro.pnpconvert
 
+import com.seriouslypro.pnpconvert.machine.DefaultMachine
+import com.seriouslypro.pnpconvert.machine.Machine
 import com.sun.xml.internal.messaging.saaj.util.ByteOutputStream
 import spock.lang.Ignore
 import spock.lang.Specification
+import spock.lang.Unroll
 
 class DPVWriterSpec extends Specification implements DPVFileAssertions {
-
 
     OutputStream outputStream
 
     List<ComponentPlacement> componentPlacements
     Components components
     Feeders feeders
-    Map<Integer, String[]> materials
+    Map<ComponentPlacement, MaterialAssignment> materialAssignments
     List<String[]> placements
     List<String[]> trays
     Optional<Panel> optionalPanel
     Optional<List<Fiducial>> optionalFiducials
     DPVHeader dpvHeader
+    BigDecimal offsetZ
+    Machine machine
 
     void setup() {
         dpvHeader = new DPVHeader(
@@ -25,7 +29,8 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
             pcbFileName: "TEST-PCB-FILE"
         )
 
-        materials = [:]
+        machine = new DefaultMachine()
+        materialAssignments = [:]
         placements = []
         trays = []
         optionalPanel = Optional.empty()
@@ -34,6 +39,8 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
         componentPlacements = []
         components = new Components()
         feeders = new Feeders()
+
+        offsetZ = 0
 
         outputStream = new ByteOutputStream()
 
@@ -45,7 +52,7 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
             DPVWriter writer = new DPVWriter()
 
         when:
-            writer.write(outputStream, dpvHeader, materials, placements, trays, optionalPanel, optionalFiducials)
+            writer.write(outputStream, machine, offsetZ, dpvHeader, materialAssignments, optionalPanel, optionalFiducials)
 
         then:
             String content = outputStream.toString()
@@ -69,7 +76,7 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
             DPVWriter writer = new DPVWriter()
 
         when:
-            writer.write(outputStream, dpvHeader, materials, placements, trays, optionalPanel, optionalFiducials)
+            writer.write(outputStream, machine, offsetZ, dpvHeader, materialAssignments, optionalPanel, optionalFiducials)
 
         then:
             String content = outputStream.toString()
@@ -88,7 +95,7 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
             optionalPanel = Optional.of(panel)
 
         when:
-            writer.write(outputStream, dpvHeader, materials, placements, trays, optionalPanel, optionalFiducials)
+            writer.write(outputStream, machine, offsetZ, dpvHeader, materialAssignments, optionalPanel, optionalFiducials)
 
         then:
             String content = outputStream.toString()
@@ -111,7 +118,7 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
             optionalFiducials = Optional.of(fiducialList)
 
         when:
-            writer.write(outputStream, dpvHeader, materials, placements, trays, optionalPanel, optionalFiducials)
+            writer.write(outputStream, machine, offsetZ, dpvHeader, materialAssignments, optionalPanel, optionalFiducials)
 
         then:
             String content = outputStream.toString()
@@ -133,6 +140,51 @@ class DPVWriterSpec extends Specification implements DPVFileAssertions {
     def 'trays should be sorted by id'() {
         expect:
             false
+    }
+
+
+    @Unroll
+    def 'placement angle - #designAngle, #pickAngle, #feederAngle'(BigDecimal designAngle, BigDecimal pickAngle, BigDecimal feederAngle, BigDecimal expectedMachineAngle) {
+
+        /*
+            Design, Feeder, Pick angles are positive clockwise, 0 to 360.
+            DipTrace EDA angles are negative clockwise, 0 to 360.  Also, Diptrace UI allows you to setting a component angle of -90, but when component is re-insepected in Diptrace UI the angle is converted to 270.
+            Machine angles are negative clockwise, -180 to +180.
+
+            Design angle = angle after conversion from EDA angle, this happens when component placements are loaded.
+            Pick angle = angle relative to pick head and feeder e.g. angle of component in tape to tape feed direction
+            Feeder angle = feeder angle relative to design
+
+            Example and Notes:
+            feeders on the left are 270 degrees out relative to the design (clockwise)
+            feeders on the right are 90 degrees out relative to the design (clockwise)
+            feeders on the left are 180 degrees out relative to the feeders on the right
+            components in a tape, e.g. usb connectors, can be different relative to the design.
+
+            IMPORTANT: Pick angles other than 0/90/180/270 will cause confusion for the vision system which will just assume
+            that the part has been picked at a strange angle and will correct to the nearest 90 degrees on the nozzle
+            during placement.
+
+            When vision is disabled the pick angle on the feeder needs to be correct.
+
+         */
+        expect:
+            expectedMachineAngle == new DPVWriter().calculateMachineAngle(designAngle, pickAngle, feederAngle)
+
+        where:
+            designAngle | pickAngle | feederAngle  | expectedMachineAngle
+            180         | 0         | 270          | 90
+            270         | 0         | 270          | 0
+            0           | 0         | 270          | -90
+            90          | 0         | 270          | 180
+            180         | 0         | 90           | -90
+            270         | 0         | 90           | 180
+            0           | 0         | 90           | 90
+            90          | 0         | 90           | 0
+            45          | 0         | 270          | -135
+            315         | 0         | 270          | -45
+            0           | 90        | 270          | 180
+            90          | 90        | 270          | 90
     }
 
 }
